Implemented initial fixes defined by PHP Stan across files.

+++ src/TwitterMapper/Object/Tweet.php
@@ -91,6 +91,6 @@ class Tweet
 
     public function getTweetAge(): int
     {
-        return ceil(((Carbon::now()->getTimestamp() - $this->getCreatedAt()->getTimestamp()) / 3600) / 24);
+        return (int) ceil(((Carbon::now()->getTimestamp() - $this->getCreatedAt()->getTimestamp()) / 3600) / 24);
     }
 }
